+++ src/app/app.component.spec.ts
@@ -1,8 +1,8 @@
 import { AppComponent } from './app.component';
 import { TestBed } from '@angular/core/testing';
-import { TestingModule } from './testing/testing.module';
-import { AuthService } from './auth/auth.service';
-import { AuthServiceFake } from './testing/testing.fakes';
+import { TestingModule } from './common/testing/testing.module';
+import { AuthService } from './common/services/auth/auth.service';
+import { AuthServiceFake } from './common/testing/testing.fakes';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
