+++ src/app/common/services/auth/auth.enum.ts
