+++ src/app/common/services/auth/auth.service.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { TestingModule } from '../testing/testing.module';
+import { TestingModule } from '../../testing/testing.module';
 
 import { AuthService } from './auth.service';
