+++ src/app/common/services/ui/ui.service.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { TestingModule } from '../testing/testing.module';
+import { TestingModule } from '../../testing/testing.module';
 
 import { UiService } from './ui.service';
