+++ src/app/common/testing/testing.fakes.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs';
-import { AuthService, IAuthStatus, IServerAuthResponse } from '../auth/auth.service'
-import { User } from '../user/user/user';
+import { AuthService, IAuthStatus, IServerAuthResponse } from '../services/auth/auth.service'
+import { User } from '../../user/user/user';
 
 export class AuthServiceFake extends AuthService {
