+++ src/app/common/testing/testing.module.ts
@@ -4,8 +4,8 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AuthService } from '../auth/auth.service';
-import { MaterialModule } from '../material/material.module';
+import { AuthService } from '../services/auth/auth.service';
+import { MaterialModule } from '../../material/material.module';
 import { AuthServiceFake } from './testing.fakes';
 
 const modules = [
