+++ src/app/free/free-routing.module.ts
@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 import { WorkoutsCalendarComponent } from './workouts/workouts-calendar/workouts-calendar.component';
 import { WorkoutsComponent } from './workouts/workouts.component';
 import { WorkoutsListComponent } from './workouts/workouts-list/workouts-list.component';
-import { AuthGuardService } from '../auth/auth-guard.service';
+import { AuthGuardService } from '../common/services/auth/auth-guard.service';
 
 const routes: Routes = [
   { path: '', redirectTo: '/free/workouts/calendar', pathMatch: 'full' },
