+++ src/app/free/workouts/edit-workout/delete-workout/delete-workout.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from 'src/app/testing/testing.module';
+import { TestingModule } from 'src/app/common/testing/testing.module';
 
 import { DeleteWorkoutComponent } from './delete-workout.component';
