+++ src/app/free/workouts/workouts-calendar/workouts-calendar.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from 'src/app/testing/testing.module';
+import { TestingModule } from 'src/app/common/testing/testing.module';
 
 import { WorkoutsCalendarComponent } from './workouts-calendar.component';
