+++ src/app/home/home.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from '../testing/testing.module';
+import { TestingModule } from '../common/testing/testing.module';
 
 import { HomeComponent } from './home.component';
