+++ src/app/page-not-found/page-not-found.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from '../testing/testing.module';
+import { TestingModule } from '../common/testing/testing.module';
 
 import { PageNotFoundComponent } from './page-not-found.component';
