+++ src/app/pro/cycles/cycles.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { EditCycleComponent } from './edit-cycle/edit-cycle.component';
 import { Cycle } from './cycle';
-import { UiService } from 'src/app/ui/ui.service';
+import { UiService } from 'src/app/common/services/ui/ui.service';
 
 @Component({
   selector: 'app-cycles',
