+++ src/app/pro/cycles/edit-cycle/delete-cycle/delete-cycle.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from 'src/app/testing/testing.module';
+import { TestingModule } from 'src/app/common/testing/testing.module';
 
 import { DeleteCycleComponent } from './delete-cycle.component';
