+++ src/app/pro/exercises/edit-exercise/delete-exercise/delete-exercise.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from 'src/app/testing/testing.module';
+import { TestingModule } from 'src/app/common/testing/testing.module';
 
 import { DeleteExerciseComponent } from './delete-exercise.component';
