+++ src/app/sign-up/sign-up.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestingModule } from '../testing/testing.module';
+import { TestingModule } from '../common/testing/testing.module';
 
 import { SignUpComponent } from './sign-up.component';
